prototype/apps/backend/onboarding/src/logic/select_career/requirements.txt
functions-framework
firebase-admin
google-cloud-firestore